add lib/buzzer e lib/rgb
