+++ lib/buzzer.h
@@ -1,0 +1,8 @@
+#ifndef BUZZER_H
+#define BUZZER_H
+
+void configurar_buzzer();
+void tocar_buzzer_alerta();
+void parar_buzzer();
+
+#endif
