+++ lib/rgb.h
@@ -1,0 +1,10 @@
+#ifndef RGB_H
+#define RGB_H
+
+#include <stdbool.h>
+
+void configurar_leds();
+void set_rgb(bool r, bool g, bool b);
+void piscar_amarelo_com_bipe();
+
+#endif
